pyproject.toml
[project]
name = "web-scrapping-alert"
version = "0.1.0"
description = ""
authors = [
    {name = "hguazza",email = "144302319+hguazza@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-telegram-bot (>=22.0,<23.0)"
]

[tool.poetry]
packages = [{include = "web_scrapping_alert", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
